Made detecting function offset in virtual function table "generic" and "clean"... relatively speaking, that is. Also reverted it to header-only.

+++ HippoMocksTest/yaffut.h
@@ -203,7 +203,7 @@ public:
     }
 
     Factory::Instance().Report ();
-    return Factory::Instance().Fail ();
+    return (int)Factory::Instance().Fail ();
   }
 };
